src/AiPipeline.TireOcr/AiPipeline.TireOcr.OcrPaddle/pyproject.toml
[project]
name = "aipipeline-tireocr-ocrpaddle"
version = "0.1.0"
description = "Ocr rest api for interacting with paddle ocr library"
authors = [
    {name = "Michal Musil"}
]
license = {text = "CC BY-NC-SA 4.0"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "aiohappyeyeballs (==2.6.1)",
    "aiohttp (==3.12.15)",
    "aiosignal (==1.4.0)",
    "aistudio-sdk (==0.3.7)",
    "annotated-types (==0.7.0)",
    "anyio (==4.10.0)",
    "astor (==0.8.1)",
    "async-timeout (==4.0.3)",
    "attrs (==25.3.0)",
    "bce-python-sdk (==0.9.46)",
    "beautifulsoup4 (==4.13.5)",
    "cachetools (==6.2.0)",
    "certifi (==2025.8.3)",
    "chardet (==5.2.0)",
    "charset-normalizer (==3.4.3)",
    "click (==8.3.0)",
    "colorlog (==6.9.0)",
    "cssselect (==1.3.0)",
    "cssutils (==2.11.1)",
    "dataclasses-json (==0.6.7)",
    "decorator (==5.2.1)",
    "distro (==1.9.0)",
    "einops (==0.8.1)",
    "et-xmlfile (==2.0.0)",
    "exceptiongroup (==1.3.0)",
    "fastapi (==0.118.2)",
    "filelock (==3.19.1)",
    "frozenlist (==1.7.0)",
    "fsspec (==2025.9.0)",
    "ftfy (==6.3.1)",
    "future (==1.0.0)",
    "h11 (==0.16.0)",
    "hf-xet (==1.1.10)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "httpx-sse (==0.4.1)",
    "huggingface-hub (==0.35.0)",
    "idna (==3.10)",
    "imagesize (==1.4.1)",
    "jinja2 (==3.1.6)",
    "jiter (==0.11.0)",
    "joblib (==1.5.2)",
    "jsonpatch (==1.33)",
    "jsonpointer (==3.0.0)",
    "langchain (==0.3.27)",
    "langchain-community (==0.3.29)",
    "langchain-core (==0.3.76)",
    "langchain-openai (==0.3.33)",
    "langchain-text-splitters (==0.3.11)",
    "langsmith (==0.4.28)",
    "lxml (==6.0.1)",
    "markupsafe (==3.0.2)",
    "marshmallow (==3.26.1)",
    "modelscope (==1.30.0)",
    "more-itertools (==10.8.0)",
    "multidict (==6.6.4)",
    "mypy-extensions (==1.1.0)",
    "networkx (==3.4.2)",
    "numpy (==2.2.4)",
    "openai (==1.108.0)",
    "opencv-contrib-python (==4.10.0.84)",
    "openpyxl (==3.1.5)",
    "opt-einsum (==3.3.0)",
    "orjson (==3.11.3)",
    "packaging (==25.0)",
    "paddleocr (==3.3.0)",
    "paddlepaddle (==3.2.0)",
    "paddlex (==3.3.3)",
    "pandas (==2.3.2)",
    "pillow (==11.1.0)",
    "premailer (==3.10.0)",
    "prettytable (==3.16.0)",
    "propcache (==0.3.2)",
    "protobuf (==6.30.1)",
    "psutil (==7.1.0)",
    "py-cpuinfo (==9.0.0)",
    "pyclipper (==1.3.0.post6)",
    "pycryptodome (==3.23.0)",
    "pydantic (==2.11.9)",
    "pydantic-settings (==2.10.1)",
    "pydantic-core (==2.33.2)",
    "pypdfium2 (==4.30.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (==1.1.1)",
    "python-multipart (==0.0.20)",
    "pytz (==2025.2)",
    "pyyaml (==6.0.2)",
    "regex (==2025.9.1)",
    "requests (==2.32.5)",
    "requests-toolbelt (==1.0.0)",
    "ruamel-yaml (==0.18.15)",
    "ruamel-yaml-clib (==0.2.12)",
    "scikit-learn (==1.7.2)",
    "scipy (==1.15.3)",
    "shapely (==2.1.1)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "soupsieve (==2.8)",
    "sqlalchemy (==2.0.43)",
    "starlette (==0.48.0)",
    "tenacity (==9.1.2)",
    "threadpoolctl (==3.6.0)",
    "tiktoken (==0.11.0)",
    "tokenizers (==0.22.0)",
    "tqdm (==4.67.1)",
    "typing-inspect (==0.9.0)",
    "typing-inspection (==0.4.1)",
    "typing-extensions (==4.15.0)",
    "tzdata (==2025.2)",
    "ujson (==5.11.0)",
    "urllib3 (==2.5.0)",
    "uvicorn (==0.37.0)",
    "wcwidth (==0.2.13)",
    "yarl (==1.20.1)",
    "zstandard (==0.25.0)"
]

[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
